Chatt Demo is an android chat application that has the added functionality of a mapping service to help suggest and arrange meeting places.
Chatt Demo is an android chat application that has the added
functionality of being a mapping service to help suggest and arrange
appropriate meeting places depending on distance from the users current
locations. You add locations of your favourite meeting places and
whenever you want to meet up with a friend appropriate locations are
suggested on a list and then shown on a map.